# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Courses-api",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/courses": {
      "post": {
        "summary": "Create course",
        "tags": [
          "courses"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 5
                  }
                },
                "required": [
                  "title"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Curso criado com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Curso criado com sucesso!",
                  "type": "object",
                  "properties": {
                    "courseId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                    }
                  },
                  "required": [
                    "courseId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all courses",
        "tags": [
          "courses"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "default": "title",
              "type": "string",
              "enum": [
                "title"
              ]
            },
            "in": "query",
            "name": "orderBy",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "type": "number"
            },
            "in": "query",
            "name": "page",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Obter todos os cursos",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Obter todos os cursos",
                  "type": "object",
                  "properties": {
                    "courses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "enrollments": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "description",
                          "enrollments"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "totalPage": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "courses",
                    "totalPage"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/courses/{courseId}": {
      "get": {
        "summary": "Get course by courseId",
        "tags": [
          "courses"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            },
            "in": "path",
            "name": "courseId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Obter curso por ID ",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Obter curso por ID ",
                  "type": "object",
                  "properties": {
                    "course": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                        },
                        "title": {
                          "type": "string"
                        },
                        "description": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "nullable": true,
                              "enum": [
                                null
                              ]
                            }
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "title",
                        "description"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "course"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Curso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Curso não encontrado"
                }
              }
            }
          }
        }
      }
    },
    "/sessions": {
      "post": {
        "summary": "Login",
        "tags": [
          "authenticate"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "token"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}